feat: all teams

+++ apps/backend/routes/api.php
@@ -121,13 +121,13 @@ Route::prefix('v1')->group(function () {
 
         Route::prefix('/team')
             ->controller(TeamController::class)
-            ->middleware('role:'.Constants::ROLE_OWNER->value.'|'.Constants::ROLE_MANAGER->value)
             ->group(function () {
                 Route::get('/', 'Index');
+                Route::get('/all', 'All');
                 Route::get('/{id}', 'Show');
-                Route::post('/', 'Create');
-                Route::put('/{id}', 'Update');
-                Route::delete('/{id}', 'Delete');
+                Route::middleware('role:'.Constants::ROLE_OWNER->value.'|'.Constants::ROLE_MANAGER->value)->post('/', 'Create');
+                Route::middleware('role:'.Constants::ROLE_OWNER->value.'|'.Constants::ROLE_MANAGER->value)->put('/{id}', 'Update');
+                Route::middleware('role:'.Constants::ROLE_OWNER->value.'|'.Constants::ROLE_MANAGER->value)->delete('/{id}', 'Delete');
             });
 
         Route::prefix('/status')
